+++ xc/programs/Xserver/lbx/lbxopts.c
@@ -300,7 +300,7 @@ ZlibParse(pno, popt, optlen, preply)
 	return (-1);
 
     pno->streamOpts.streamCompInit = ZlibInit;
-    pno->streamOpts.streamCompArg = (pointer) level;
+    pno->streamOpts.streamCompArg = (pointer)(long)level;
     pno->streamOpts.streamCompStuffInput = ZlibStuffInput;
     pno->streamOpts.streamCompInputAvail = ZlibInputAvail;
     pno->streamOpts.streamCompFlush = ZlibFlush;
